# Admin/src/app/services/models/application-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApplicationRequest {
  academicLevel: string;
  amount: number;
  applicationRegion: string;
  dateOfBirth: string;
  departmentOfOrigin: string;
  email: string;
  examType: string;
  financialRessource: string;
  formationMode: string;
  freeCandidate: boolean;
  learningLanguage: string;
  matrimonialSituation: string;
  nationIdNumber: string;
  nationality: string;
  numberOfKid: number;
  paymentMethod: string;
  placeOfBirth: string;
  regionOrigins: string;
  repeatCandidate: boolean;
  secretCode: number;
  sessionYear?: string;
  sex: string;
  speciality?: string;
  trainingCenterAcronym: string;
}