# Admin/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specifiction - APPLYCAM",
    "description": "OpenApi documentation for applycam project",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "jotsamikael",
      "url": "https://dodibo.com/",
      "email": "jotsamikael@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://someurl.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV."
    },
    {
      "url": "https://something:8088/api/v1",
      "description": "Prod ENV."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/application/reject/{id}": {
      "post": {
        "tags": [
          "application"
        ],
        "operationId": "rejectApplication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "comment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/application/PersonalInformation": {
      "post": {
        "tags": [
          "application"
        ],
        "operationId": "candidateAppliance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/application/validate/{id}": {
      "patch": {
        "tags": [
          "application"
        ],
        "operationId": "validateApplication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/application/delete/{applicationId}": {
      "patch": {
        "tags": [
          "application"
        ],
        "operationId": "deleteApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/application/PersonalInformation/documents": {
      "patch": {
        "tags": [
          "application"
        ],
        "operationId": "uploadCandidateFile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "birthCertificate",
                  "cniFile",
                  "diplomFile",
                  "photo"
                ],
                "type": "object",
                "properties": {
                  "cniFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "birthCertificate": {
                    "type": "string",
                    "format": "binary"
                  },
                  "diplomFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "oldApplyanceFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "stageCertificate": {
                    "type": "string",
                    "format": "binary"
                  },
                  "cv": {
                    "type": "string",
                    "format": "binary"
                  },
                  "financialJustification": {
                    "type": "string",
                    "format": "binary"
                  },
                  "letter": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/application/get-application-By-candidate": {
      "get": {
        "tags": [
          "application"
        ],
        "operationId": "searchApplicationsByCandidateName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/application/get-all": {
      "get": {
        "tags": [
          "application"
        ],
        "operationId": "getAllApplications",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationRequest": {
        "required": [
          "academicLevel",
          "amount",
          "applicationRegion",
          "dateOfBirth",
          "departmentOfOrigin",
          "email",
          "examType",
          "financialRessource",
          "formationMode",
          "freeCandidate",
          "learningLanguage",
          "matrimonialSituation",
          "nationIdNumber",
          "nationality",
          "numberOfKid",
          "paymentMethod",
          "placeOfBirth",
          "regionOrigins",
          "repeatCandidate",
          "secretCode",
          "sex",
          "trainingCenterAcronym"
        ],
        "type": "object",
        "properties": {
          "sex": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nationIdNumber": {
            "type": "string"
          },
          "academicLevel": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          },
          "examType": {
            "type": "string"
          },
          "departmentOfOrigin": {
            "type": "string"
          },
          "applicationRegion": {
            "type": "string"
          },
          "regionOrigins": {
            "type": "string"
          },
          "trainingCenterAcronym": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "freeCandidate": {
            "type": "boolean"
          },
          "repeatCandidate": {
            "type": "boolean"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "paymentMethod": {
            "type": "string"
          },
          "secretCode": {
            "type": "integer",
            "format": "int64"
          },
          "matrimonialSituation": {
            "type": "string"
          },
          "learningLanguage": {
            "type": "string"
          },
          "formationMode": {
            "type": "string"
          },
          "financialRessource": {
            "type": "string"
          },
          "placeOfBirth": {
            "type": "string"
          },
          "numberOfKid": {
            "type": "integer",
            "format": "int32"
          },
          "sessionYear": {
            "type": "string"
          },
          "speciality": {
            "type": "string"
          }
        }
      },
...
      "ApplicationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "candidateName": {
            "type": "string"
          },
          "speciality": {
            "type": "string"
          },
          "applicationRegion": {
            "type": "string"
          },
          "applicationYear": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PageResponseApplicationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}